pyproject.toml
[build-system]
build-backend = "uv_build"
requires = ["uv_build>=0.9.11,<0.10.0"]

[project]
name = "litestar-flags"
version = "0.1.0"
description = "Production-ready feature flags library for Litestar"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Jacob Coffee", email = "jacob@z7x.org" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed",
]
keywords = ["litestar", "feature-flags", "feature-toggles", "a/b-testing", "rollouts"]
dependencies = [
    "litestar>=2.0.0",
]

[project.optional-dependencies]
database = [
    "advanced-alchemy>=0.10.0",
    "sqlalchemy[asyncio]>=2.0.0",
]
redis = [
    "redis>=5.0.0",
]
otel = [
    "opentelemetry-api>=1.20.0",
]
logging = [
    "structlog>=24.1.0",
]
observability = [
    "litestar-flags[otel,logging]",
]
workflows = [
    "litestar-workflows>=0.1.0",
]
openfeature = [
    "openfeature-sdk>=0.8.0",
]
all = [
    "litestar-flags[database,redis,observability,workflows,openfeature]",
]

[project.urls]
Homepage = "https://github.com/JacobCoffee/litestar-flags"
Documentation = "https://github.com/JacobCoffee/litestar-flags#readme"
Repository = "https://github.com/JacobCoffee/litestar-flags"
Issues = "https://github.com/JacobCoffee/litestar-flags/issues"

[project.entry-points."litestar.plugins"]
feature_flags = "litestar_flags:FeatureFlagsPlugin"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "A",    # flake8-builtins
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "D",    # pydocstyle
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "Q",    # flake8-quotes
    "RUF",  # Ruff-specific rules
    "S",    # flake8-bandit
    "T20",  # flake8-print
    "UP",   # pyupgrade
    "W",    # pycodestyle warnings
]
ignore = [
    "D100",   # Missing docstring in public module
    "D104",   # Missing docstring in public package
    "D105",   # Missing docstring in magic method (__repr__, etc.)
    "D107",   # Missing docstring in __init__
    "D203",   # 1 blank line required before class docstring
    "D213",   # Multi-line docstring summary should start at the second line
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["S101", "S105", "D", "B017", "C401"]
"benchmarks/**/*" = ["S101", "D", "T201"]

[tool.ty]

[tool.ty.environment]
python = ".venv"

[tool.ty.rules]
deprecated = "warn"

[tool.ty.src]
exclude = ["docs/conf.py", "tests/", "src/litestar_flags/storage/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra -q --cov=litestar_flags --cov-report=term-missing"

[tool.coverage.run]
source = ["src/litestar_flags"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]

[dependency-groups]
docs = [
    "sphinx>=7.0.0",
    "shibuya>=2024.8.0",
    "sphinx-autobuild>=2024.0.0",
    "sphinx-autodoc-typehints>=2.0.0",
    "sphinx-copybutton>=0.5.0",
    "sphinx-design>=0.5.0",
    "myst-parser>=2.0.0",
    "uuid-utils>=0.6.0",  # Suppress advanced-alchemy warnings
]
lint = [
    "ty>=0.0.1a27",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
]
test = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-benchmark>=5.1.0",
    "httpx>=0.27.0",
    "polyfactory>=2.15.0",
    # Storage backends for testing
    "advanced-alchemy>=0.10.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
    "redis>=5.0.0",
    "fakeredis>=2.21.0",
]
dev = [
    {include-group = "docs"},
    {include-group = "lint"},
    {include-group = "test"},
]

[tool.uv]
default-groups = ["dev"]

[tool.uv.sources]
litestar-workflows = { path = "../litestar-workflows" }
